pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ask-dcnc"
version = "1.4.1"
description = "A chatbot that can answer questions about RMIT programs and courses using official program plans and course guides."
readme = "readme.md"
requires-python = ">=3.10"

dependencies = [
    "streamlit~=1.45.1",
    "boto3~=1.38.26",
    "python-dotenv~=1.1.0",
    "langchain~=0.3.21",
    "langchain_aws~=0.2.24",
    "langchain-community~=0.3.24",
    "toml~=0.10.2",
    "langgraph~=0.4.8",
    "loguru~=0.7.3",
    "beautifulsoup4~=4.13.4"
]

[tool.setuptools]
packages = ["ask_dcnc"]
